+++ packages/openai/src/openai-provider.ts
@@ -4,7 +4,7 @@ import {
   LanguageModelV3,
   ProviderV3,
   SpeechModelV3,
-  TranscriptionModelV2,
+  TranscriptionModelV3,
 } from '@ai-sdk/provider';
 import {
   FetchFunction,
@@ -81,7 +81,7 @@ Creates a model for image generation.
   /**
 Creates a model for transcription.
    */
-  transcription(modelId: OpenAITranscriptionModelId): TranscriptionModelV2;
+  transcription(modelId: OpenAITranscriptionModelId): TranscriptionModelV3;
 
   /**
 Creates a model for speech generation.